refactor: 모든 컴포넌트의 기반이 되는 Components.js를 상위 계층으로 폴더구조 변경
