+++ src/step2/components/@common/Button.js
@@ -1,4 +1,4 @@
-import Component from "../Component";
+import Component from "../../core/Component";
 
 export default class Button extends Component {
   template() {
