+++ src/step2/components/AmountInput.js
@@ -2,7 +2,7 @@ import { INPUT_MESSAGES } from "../lib/constants.js";
 import LottoPurchase from "../service/LottoPurchase.js";
 import { qs } from "../../utils/domHelper.js";
 import Button from "./@common/Button.js";
-import Component from "./Component.js";
+import Component from "../core/Component.js";
 import Validator from "../validator/Validator.js";
 
 export default class AmountInput extends Component {
