+++ src/step2/components/StatisticsModal.js
@@ -1,7 +1,7 @@
 import { LOTTO_RANK } from "../lib/constants";
 import { qs } from "../../utils/domHelper";
 import Button from "./@common/Button";
-import Component from "./Component";
+import Component from "../core/Component";
 
 export default class StatisticsModal extends Component {
   template() {
