+++ src/step2/components/UserInput.js
@@ -2,7 +2,7 @@ import LottoResult from "../service/LottoResult";
 import { qs, qsAll } from "../../utils/domHelper";
 import Validator from "../validator/Validator";
 import Button from "./@common/Button";
-import Component from "./Component";
+import Component from "../core/Component";
 
 export default class UserInput extends Component {
   constructor(element, props) {
