+++ src/step2/core/Component.js
